requirements.txt
faiss-cpu==1.8.0.post1
FlagEmbedding==1.3.2
numpy==1.26.4
tqdm==4.66.1
